added menu but crashes
