+++ app/build.gradle
@@ -15,6 +15,9 @@ android {
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_1_8.toString()
+    }
     buildTypes {
         release {
             minifyEnabled false
@@ -35,4 +38,6 @@ dependencies {
     implementation 'com.google.android.material:material:1.0.0'
     implementation 'org.nanohttpd:nanohttpd:2.3.1'
     implementation 'com.android.volley:volley:1.1.1'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.2.1'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.2.1'
 }
