+++ app/src/main/res/menu/upper_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/dialog_create"
+        android:icon="@drawable/ic_plus"
+        android:menuCategory="secondary"
+        android:title="@string/new_dialog" />
+</menu>
