+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@id/mainActivity">
+    <activity
+        android:id="@+id/mainActivity"
+        android:name="com.example.jam.MainActivity"
+        android:label="activity_main"
+        tools:layout="@layout/activity_main" >
+        <action
+            android:id="@+id/to_dialog"
+            app:destination="@id/dialogActivity"
+            app:enterAnim="@anim/fade_in"
+            app:exitAnim="@anim/fade_out"
+            app:popEnterAnim="@anim/fade_in"
+            app:popExitAnim="@anim/fade_out" />
+    </activity>
+    <activity
+        android:id="@+id/dialogActivity"
+        android:name="com.example.jam.DialogActivity"
+        android:label="dialog_activity"
+        tools:layout="@layout/dialog_activity" />
+    <dialog
+        android:id="@+id/dialog_create"
+        android:name="com.example.jam.CreateNewDialogFragment"
+        android:label="create_new_dialog_fragment"
+        tools:layout="@layout/create_new_dialog_fragment" >
+        <action
+            android:id="@+id/action_back_to_main"
+            app:destination="@id/mainActivity" />
+    </dialog>
+</navigation>
