+++ app/src/main/res/values/anim/slide_in_right.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <translate android:fromXDelta="100%" android:toXDelta="0%"
+               android:fromYDelta="0%" android:toYDelta="0%"
+               android:duration="700"/>
+</set>
