+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,20 @@
     <string name="ipHint">IP</string>
     <string name="messageHint">Message</string>
     <string name="ipFish">ip</string>
+    <string name="settings">Settings</string>
+    <string name="home">Home</string>
+    <string name="deeplink">Deep Link</string>
+    <string name="step_one">Step One</string>
+    <string name="step_two">Step Two</string>
+    <string name="deep_link_instructions">Type your deep link args</string>
+    <string name="settings_fragment_text">Settings Fragment</string>
+    <string name="shopping_fragment_text">Shopping Fragment</string>
+    <string name="navigate_destination_button">Navigate to destination</string>
+    <string name="navigate_action_button">Navigate with action</string>
+    <string name="next_step_button">Navigate next step</string>
+    <string name="finish_button">Finish Flow</string>
+    <string name="send_deep_link_notification_button">Send deep link notification</string>
+    <string name="new_dialog">Create new dialog</string>
+    <string name="confirm_button">Confirm</string>
+    <string name="nameHint">Name</string>
 </resources>
